[Module_1][Update Fix PR Review from Hermawan day 3 v3]
